+++ packages/@expo/cli/src/start/server/metro/withMetroMultiPlatform.ts
@@ -562,6 +562,10 @@ export async function withMetroMultiPlatformAsync(
     }
     // @ts-expect-error: watchFolders is readonly
     config.watchFolders.push(path.join(require.resolve('metro-runtime/package.json'), '../..'));
+    if (isReactCanaryEnabled) {
+      // @ts-expect-error: watchFolders is readonly
+      config.watchFolders.push(path.join(require.resolve('@expo/cli/package.json'), '..'));
+    }
   }
 
   // @ts-expect-error