+++ packages/@expo/metro-config/src/ExpoMetroConfig.ts
@@ -254,6 +254,7 @@ export function getDefaultConfig(
     transformer: {
       // Custom: These are passed to `getCacheKey` and ensure invalidation when the version changes.
       // @ts-expect-error: not on type.
+      unstable_renameRequire: false,
       postcssHash: getPostcssConfigHash(projectRoot),
       browserslistHash: pkg.browserslist
         ? stableHash(JSON.stringify(pkg.browserslist)).toString('hex')