+++ packages/expo-updates/e2e/setup/project.ts
@@ -51,6 +51,7 @@ function getExpoDependencyChunks({
             'expo-localization',
             'expo-crypto',
             'expo-network',
+            'expo-video',
           ],
         ]
       : []),
@@ -341,8 +342,8 @@ async function preparePackageJson(
       ...packageJson,
       dependencies: {
         ...packageJson.dependencies,
-        'react-native': 'npm:react-native-tvos@~0.74.1-0',
-        '@react-native-tvos/config-tv': '^0.0.9',
+        'react-native': 'npm:react-native-tvos@~0.74.2-0',
+        '@react-native-tvos/config-tv': '^0.0.10',
       },
       expo: {
         install: {